Updates to pathfinding
Been playing around with pathfinding for a few days; the implementation of paththfinding.path has been improved.
The procedural_navmesh still needs work; whilst within a particular navmesh paths can be constructed and obstacles
moved, inter-navmesh pathing is still proving to be diffucult. This is in part due to the fact that I am relying
on a perfect hashing function existing for the points to be pathed over. Such a function does not exist when we
start looking at arbitrarily-sized pathfinding regions. I kind of want to move away from this towards a more linked-list
like approach, especially if such an arangement makes grid optimization more tractable.

The current plan is to design a few characters, using pathfinding.path for now, to get a better idea of what the
pathfinding needs to be capable of, before further attacking the navmesh problem.