Initial player networking
Added various utilities to the networking enginge, including option
to determine if a networked object was created on the currently
running client and options to control object serialization. This was
useful in the player case as the client that created the player should
have full control over that player, which should not be recciving
serialization updates from other clients.

Also added a utility to simulate ping from the client to the server,
as well as message queues on the client and server to combine messages
into single stream writes.