Added networked characters/spawners
Added networked_object_spawner, a type of world_object that can
spawn a networked_object/ensures that it is only created once by
waiting until it is sure that it was not created elsewhere.

Clients now have unique authority over networked objects to allow
clients to decide if they are in control of NPCs etc.

Networked variables can now tell if they are being deseiralized
from the network for the first time, to allow first-time initialization.
This is used by all networked objects to ensure that the positions loaded
for the first time (i.e the saved positions from last time) are applied
directly (rather than LERPed).