Debugged disonnect/reconnect
Clicking on save + quit now takes you to the main menu. This allows
you to immediately reconnect from the same game instance, and, as a
result, static variables persist. This caused a variety of bugs, which
have now been fixed (hopefully). Also generally tidied up networking
code in the process.