Fixed item dupe bug
When two clients simultaneously deleted an object. They would both end
up with a copy of the object. Now, the delete() function accepts an
argument for a callback to be called when the server successfully
deletes the object. As a result, only the first client to delete the
object will call the callback. By adding the deleted item to the
inventory only in this callback, the bug is fixed.