Made inventory a networked object
Inventories are now in-game objects which reference a specific ui object
that contains the inventory slots that make up the inventory. This is so
they can be loaded/unloaded in the same way as other networked objects,
with proximity tests etc. In the process, inventory_section has now been
removed; seperate inventory sections should now either be seperate ui
objects belonging to seperate inventories, or the neccassary difference
in behaviour should be achived by inheriting from inventory_slot.