Updates to item network
Removed experimental item_transport_node. Implemented item lift.
Networked item flow in gutters, so that item flow from outside
the loaded area of the map is possible (still has bugs, such as when
a player deletes a gutter, the next gutter down the line thinks it
needs to mimic the item flow as if it is at the edge of the loaded
world).