Fixed networking bug when world is large
When the world was large enough so that loading it required more data
than the TCP read buffer size (64K in my test case), messages at the end
of the buffer would be truncated (and anything after that would
therefore be read in as garbage). Message truncation is now detected and
the first part is stored for later re-assembely with it's tail (in both
the server/client code).