Fixed bug where equipped item wasn't networked properly
In the process, an issue with networked variables was identified,
whereby networked_variables within collections (such as networked_list<T> where
T:networked_variable) sent their own updates (rather than relying on the
collection to do this for them). A hotfix has been introduced with a
send_updates flag on networked variables, which is set to false inside
networked_lists etc. However, it is still possible to run into this bug
quite easily when initializing networked collections. For example:

// Owned by a networked object
list = new networked_list<net_int>()

// Not owned by a networked object (as it will be
// used wihtin a networked list)
var a = new net_int();

a.value = 1; // Will queue an update
l.add(a); // Will set a.send_update = false, but update already queued!

Switching the order of the last two lines fixes this issue in the above
code. Either way, the list will send updates correctly (which is why
this bug has gone un-noticed until now).

A catch for code like the above has been added, whereby an error will be
thrown if a networked_variable without an owner ends up in the
queued_variable_updates queue (checked when de-queued).

However, it is often difficult to
determine from this error where/when the update was queued. This would
be easier if the error was thrown at queue time, but this
cant happen because it is possible that the list in the above code
snippet is yet to obtain an owner, but will do shortly (this is common
in the codebase because the owner is set just *after*
on_init_networked_variables is called)

I will probably open a branch to investigate cleaner solutions.
This might consist of passing an owner to the constructor of
networked_variables (which might also be used to build the
networked_variable list of the owner). However, this is quite
a big change to the signature of networked variables and will
require correspondingly large changes to the codebase.