Fixed duplicate manager objects
Both those created when you make a new world, and those loaded from the
startup map file were persisting. The oldest ones (typically, these will
be from the startup map) are now kept, and the others deleted.

Also added a clear world command to create test worlds with nothing in
them.

Notes:
If the managers in the startup file get corrupted somehow, then it could
be difficult to fix them?