Optimized attacker entrypoint
This was done by taking advantage of the geometry update listener to
trigger repathing/path validation (rather than just having this on a
timer).

Also fixed a bug in random paths, where the goal point was added during
path reconstruction (this is not acceptable, as random paths don't have
a useful definition of a goal point - it usually just set to the start
point).